# Problem: unnecessary GetBalance in ante handlers