@yihuang 2024-08-16
again, does this optimisation affects benchmark result, I guess the effect of this is even smaller than the account PR, which is almost unnoticable in the benchmark result.