@mmsqe 2024-08-16
> again, how does this affect the benchmark result, I guess the effect of this is even smaller than the account PR, which is almost unnoticeable in the benchmark result.

Do you mean GetBalance is faster than GetAccount, no need to cache?