@yihuang 2024-08-16
> > again, how does this affect the benchmark result, I guess the effect of this is even smaller than the account PR, which is almost unnoticeable in the benchmark result.
> 
> Do you mean GetBalance is faster than GetAccount, no need to cache?

what do you think we are caching here? the raw bytes fetched from db is already cached by cachekv store, so what we are caching here is just to avoid the decoding step, if the decoding speed is similar to or even faster than the added map operations, then there's no point to add this cache, right?